LPS-65385 Move DestinationNames to com.liferay.push.notifications.constants
