+++ modules/apps/push-notifications/push-notifications-sender/push-notifications-sender-android/src/main/java/com/liferay/push/notifications/sender/android/internal/AndroidPushNotificationsSender.java
@@ -28,8 +28,8 @@ import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.messaging.MessageBusUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.push.notifications.constants.PushNotificationsConstants;
+import com.liferay.push.notifications.constants.PushNotificationsDestinationNames;
 import com.liferay.push.notifications.exception.PushNotificationsException;
-import com.liferay.push.notifications.messaging.DestinationNames;
 import com.liferay.push.notifications.sender.PushNotificationsSender;
 import com.liferay.push.notifications.sender.Response;
 import com.liferay.push.notifications.sender.android.internal.configuration.AndroidPushNotificationsSenderConfiguration;
@@ -131,7 +131,8 @@ public class AndroidPushNotificationsSender implements PushNotificationsSender {
 				result, token, payloadJSONObject);
 
 			MessageBusUtil.sendMessage(
-				DestinationNames.PUSH_NOTIFICATION_RESPONSE, response);
+				PushNotificationsDestinationNames.PUSH_NOTIFICATION_RESPONSE,
+				response);
 
 			if ((multicastResult.getCanonicalIds() == 0) &&
 				(multicastResult.getFailure() == 0)) {
